Extract version assertion function
