+++ waspc/tools/utils.ts
@@ -4,6 +4,7 @@ import { execFileSync, type ExecFileSyncOptions } from "node:child_process";
 import { existsSync, readdirSync, readFileSync } from "node:fs";
 import { join } from "node:path";
 import { fileURLToPath } from "node:url";
+import { getWaspcVersion } from "./get-waspc-version.ts";
 
 export function getRepoRootPath(): string {
   return fileURLToPath(new URL("../..", import.meta.url));
@@ -26,6 +27,23 @@ export function getPackageJson(dir: string): {
 } {
   const packageJsonPath = join(dir, "package.json");
   return JSON.parse(readFileSync(packageJsonPath, "utf-8"));
+}
+
+export function assertPackageVersionMatchesWaspc(
+  packageName: string,
+  packageVersion: string,
+): void {
+  const waspcVersion = getWaspcVersion();
+
+  if (packageVersion !== waspcVersion) {
+    console.error(
+      `ERROR: ${packageName} package version (${packageVersion}) != current Wasp version (${waspcVersion}).`,
+    );
+    console.error(
+      `       Update the package version in package.json to ${waspcVersion}.`,
+    );
+    throw new Error();
+  }
 }
 
 export function runCmd(
